+++ test/unit/vagrant/devcommands/internal_spec.rb
@@ -10,16 +10,16 @@ describe VagrantPlugins::DevCommands::Internal do
 
       internal = described_class.new(reg)
 
-      expect { internal.run('help') }.to output(/version/).to_stdout
-      expect { internal.run('help') }.to output(/help/).to_stdout
+      expect { internal.run('help', []) }.to output(/version/).to_stdout
+      expect { internal.run('help', []) }.to output(/help/).to_stdout
     end
   end
 
   describe 'internal help command' do
-    it 'notifies if no command is known' do
+    it 'notifies if no command is available' do
       internal = described_class.new(registry.new)
 
-      expect { internal.run('help') }.to output(/no commands/i).to_stdout
+      expect { internal.run('help', []) }.to output(/no commands/i).to_stdout
     end
 
     it 'displays a list of available commands' do
@@ -31,8 +31,8 @@ describe VagrantPlugins::DevCommands::Internal do
 
       internal = described_class.new(reg)
 
-      expect { internal.run('help') }.to output(/bar/).to_stdout
-      expect { internal.run('help') }.to output(/foo/).to_stdout
+      expect { internal.run('help', []) }.to output(/bar/).to_stdout
+      expect { internal.run('help', []) }.to output(/foo/).to_stdout
     end
 
     it 'display description if available' do
@@ -44,7 +44,9 @@ describe VagrantPlugins::DevCommands::Internal do
 
       internal = described_class.new(reg)
 
-      expect { internal.run('help') }.to output(/has a description/).to_stdout
+      expect { internal.run('help', []) }.to(
+        output(/has a description/).to_stdout
+      )
     end
   end
 
@@ -53,7 +55,7 @@ describe VagrantPlugins::DevCommands::Internal do
       version  = VagrantPlugins::DevCommands::VERSION
       internal = described_class.new(registry.new)
 
-      expect { internal.run('version') }.to output(/#{version}/).to_stdout
+      expect { internal.run('version', []) }.to output(/#{version}/).to_stdout
     end
   end
 end